% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_ref}
\alias{add_ref}
\title{Add a Reference to a Bibliography File}
\usage{
add_ref(bib_path)
}
\arguments{
\item{bib_path}{full path to bibliography file to write to}
}
\value{
Nothing
}
\description{
'add_ref()'  Reads clipboard contents and writes it to references file
}
