% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_square_brackets.R
\name{ci_square_brackets}
\alias{ci_square_brackets}
\title{Title Change CI Formating, when in Round Brackets}
\usage{
ci_square_brackets(CI)
}
\arguments{
\item{CI}{fomrated confidence interval from format_num_denom_accuracy_measures()}
}
\value{
string: fomrated confidence interval
}
\description{
When writing manuscript, default formating of CI is round brackets, if mentioned
in round brackets, then changes theround brackets to square ones
}
