% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_vect_2_text.R
\name{collapse_vect_2_text}
\alias{collapse_vect_2_text}
\title{Takes a named numeric vector vector and collapses it into publication style sentences.}
\usage{
collapse_vect_2_text(vect, plural_text_unit, singular_text_unit = NULL)
}
\arguments{
\item{vect}{String: named numeric vector.}

\item{plural_text_unit}{String: Noun that is being counted in plural. If the plural does not in 's'
the automatic singular plural changes will not work with a warning}

\item{singular_text_unit}{default is NULL, needed if numeric vectors contains 1
and plural form does not end with s}
}
\value{
text sentence, should be gramatically formated for publication style
}
\description{
This function takes a numeric vector and spells it out in a sentence. numerics < 10 will be spelled out
singular and plural works somewhat, anot all cases are covered
}
