% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_lab_pages.R
\name{combine_lab_pages}
\alias{combine_lab_pages}
\title{Title}
\usage{
combine_lab_pages(lab_list)
}
\arguments{
\item{lab_list}{list of parameters produced by prep_labor_pdf_pages}
}
\value{
named list of labor parameters, containing date_time and value
}
\description{
Title
}
