% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num_denom_accuracy_measures.R
\name{format_num_denom_accuracy_measures}
\alias{format_num_denom_accuracy_measures}
\title{Title Format summary stats into CI string for Publication}
\usage{
format_num_denom_accuracy_measures(tib, output = "both", round_digs = 2)
}
\arguments{
\item{tib}{Tibble from num_denom_accuracy_measures()}

\item{output}{String, c("value_only", "num_denom", "ci", "both")}

\item{round_digs}{Int: number of digits to round to}
}
\value{
String fomrated ready for Publication containing acc, sens and spec
with/without CI and or numerator and denominator
}
\description{
takes the tibble produced by num_denom_accuracy_measures()
Output types
value_only: returns the Accuracy, sensitivity and spec only
num_denom: returns numerator and denominator only
ci: returns CI only
both: (Default) Returns both num/denum and CI
}
