% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_2_words.R
\name{num_2_words}
\alias{num_2_words}
\title{Changing numbers to Written letters
In formal writing numbers at the beginning of sentences showed be written in letters.
This functions does this. I claim no credit. Credits are in comments in the source code}
\usage{
num_2_words(x)
}
\arguments{
\item{x}{A number}
}
\value{
String with the number x written out
}
\description{
Changing numbers to Written letters
In formal writing numbers at the beginning of sentences showed be written in letters.
This functions does this. I claim no credit. Credits are in comments in the source code
}
\examples{
num_2_words(5)
}
