% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_denom_accuracy_measures.R
\name{num_denom_accuracy_measures}
\alias{num_denom_accuracy_measures}
\title{Title Calculate Numerator and Denom, CI from Confusion Matrix}
\usage{
num_denom_accuracy_measures(cm, positive_class)
}
\arguments{
\item{cm}{confusion matrix from Tidy models}

\item{positive_class}{String defining the positive class}
}
\value{
Tibble containing numerator, denom, CI and summary stats
}
\description{
Title Calculate Numerator and Denom, CI from Confusion Matrix
}
