% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_in_browser.R
\name{ref_in_browser}
\alias{ref_in_browser}
\title{Load Reference in Browser}
\usage{
ref_in_browser(
  wanted_ref,
  bib_path,
  browser_path =
    "C:\\\\Users\\\\elshikh\\\\AppData\\\\Roaming\\\\Microsoft\\\\Windows\\\\Start Menu\\\\Programs\\\\UKL Firefox"
)
}
\arguments{
\item{wanted_ref}{reference name as saved in bibtex file or as would be
referenced in quarto document}

\item{bib_path}{full path to bibliography file to write to}

\item{browser_path}{path to browser to load the wanted reference}
}
\value{
Nothing
}
\description{
'ref_in_browser()' searches for URL in the bibtex reference and lods it in
selected browser
}
