% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_nested_list.R
\name{rename_nested_list}
\alias{rename_nested_list}
\title{Title replace pattern in nested list}
\usage{
rename_nested_list(nested_list, pattern, replacement)
}
\arguments{
\item{nested_list}{a nested named list to rename. behaviour with unnamed lists is not tested}

\item{pattern}{as text string}

\item{replacement}{as text string}
}
\value{
list
}
\description{
Recursively applies str_replace to the names of all elements of a named nested list
credit to https://stackoverflow.com/a/63075776/14604591 and to https://stackoverflow.com/a/38539734/14604591
}
