% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_text.R
\name{render_text}
\alias{render_text}
\title{Title Render quarto or Rmarkdown text to console}
\usage{
render_text(tex = NULL)
}
\arguments{
\item{tex}{String: to render, or default: read text from clipboard}
}
\value{
Nothing
}
\description{
When writing a quarto or Rmarkdown document, this function is
useful to check the rendered output, e. g. checking the variables
are correct and are rendered correctly.
BUG!!!: sometimes extra spaces are in the output
}
