% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_text_to_vector.R
\name{locate_text_to_vector}
\alias{locate_text_to_vector}
\title{Locate Pattern in Character Columns within Mutate}
\usage{
locate_text_to_vector(col, pattern, back_window = 10, front_window = 80)
}
\arguments{
\item{col}{column name unquoted}

\item{pattern}{text un quotes}

\item{back_window}{number of characters before search term to include in output}

\item{front_window}{number of characters after search term to include in output}
}
\value{
vector of strings
}
\description{
Tibble mut grouped first using rowwise()
Used within mutate(). Takes column name (unquoted) and text (in quotes),
searches for text in this column, and returns its location. Then text is
extracted and returned as a charachter vector
Mutate call has to be within a list, that is later unnet
}
\examples{
df <- tibble::tibble(col_id = letters[1:2], col_text = c("I love ananas", "I love lololo"))
df |> dplyr::rowwise() |> dplyr::mutate(lo = list(locate_text_to_vector(col_text, "lo", 2, 7)))
}
