% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_labor_pdfs.R
\name{prep_labor_pdf_pages}
\alias{prep_labor_pdf_pages}
\title{Title Create tidy table from labor PDF}
\usage{
prep_labor_pdf_pages(
  pdf_df,
  relevant_filter,
  empty_filter,
  wanted_labs,
  y_upper_limit,
  stop_term,
  verbose = TRUE
)
}
\arguments{
\item{pdf_df}{a single dataframe from the list produced by pdftools::pdf_data()}

\item{relevant_filter}{stringr::regex() expression containing text to search for to define the where the columns along the x axis}

\item{empty_filter}{stringr::regex() expression containing string patterns that will be changed to NA}

\item{wanted_labs}{stringr::regex() expression containing labor parameters needed}

\item{y_upper_limit}{stringr::regex() expression for upper limit to include (usually name of the first column)}

\item{stop_term}{stringr::regex() expression to stop at if encountered}

\item{verbose}{prints piz and page number of the dicument being processed to console}
}
\value{
a tibble containing date_time and value
}
\description{
Title Create tidy table from labor PDF
}
