% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_nested_lists.R
\name{tidy_nested_lists}
\alias{tidy_nested_lists}
\title{Title}
\usage{
tidy_nested_lists(nested_list)
}
\arguments{
\item{nested_list}{complex nested list, tested with sublists or subtibbles}
}
\value{
tidy tibble. column names are numbered 0 based, to correspond to the depth argument in maap_depth()
}
\description{
Title
}
